Navigate to most probable syllable after the initial guess
Turns out it's impossible (at least we don't know how) to guess the
most probable syllable. So we make an initial effort (our best guess),
and then check if there are permutations of symbols that give a more
probable word. And so on until we reached "the top of the hill" - the
word that's most probable.